Added property-based bootstrap for users of xml-based MyBatis bootstrap and added stub for an integration test.
Integration tests should be moved to a submodules, so they don't pollute
the main code base... having them in isolation should also help to notice
when we're doing something stupid that breaks the users.

Property-based bootstrap is to be used when no dependency injection is
available and when user doesn't construct MyBatis' configuration by hand,
from Java code... so it will cover A LOT of users.

This module reads system property 'metrics.registry.provider' that should
contain fully qualified class name of an implementation of
javax.inject.Provider<MetricRegistry>. This implementation is instanitated
using no-arg constructor and is used to, suprise suprise, get
MetricRegistry instance and use that to initialize InstrumentingInterceptor.